+++ server/routes/api/v1/dbinsert.js
@@ -1,8 +1,8 @@
 const { Client } = require('pg');
+const config = require('config');
 const client = new Client({
-  user: 'postgres',
-  password: 'Alvaro334',
-  // host: 'Ojumo',
+  user: config.get('userName'),
+  password: config.get('myPass'),
   port: 5432,
   database: 'test'
 });
@@ -14,11 +14,6 @@ const connectDB = async () => {
 const insertUser = async (name, email, avatar, password) => {
   const table = 'users';
   try {
-    /* 
-    const result = await client.query(queryText, ['841l14yah', 'test@te.st']);
-  const newlyCreatedUserId = result.rows[0].id;
-    
-    */
     await client.query('BEGIN');
     const sql = `INSERT INTO ${table}(name,email,avatar,password) VALUES ($1, $2,$3,$4) RETURNING id`;
     const params = [name, email, avatar, password];
@@ -44,13 +39,23 @@ const querydb = async qbody => {
     const sql =
       ' SELECT id, name, email, avatar, is_admin, created_date, modified_date FROM users WHERE id = $1';
     const params = [qbody];
-    console.log(await client.query(sql, params));
-    return await client.query(sql, params);
-    // const result = await client.query(sql, params);
-    // const newlyCreatedUserId = result.rows[0].id;
+    const result = await client.query(sql, params);
+    const newlyCreatedUserId = result.rows[0];
+
     return newlyCreatedUserId;
   } catch (error) {
     return error;
   }
 };
-module.exports = { insertUser, connectDB, querydb };
+
+const qpass = async qbody => {
+  try {
+    const sql = 'SELECT id, password FROM users WHERE email = $1';
+    const params = [qbody];
+    const result = await client.query(sql, params);
+    // console.log(result.rows);
+
+    return result.rows[0];
+  } catch (error) {}
+};
+module.exports = { insertUser, connectDB, querydb, qpass };
